+++ packages/core/src/lib/implementation/runner.spec.ts
@@ -1,7 +1,8 @@
 import { describe, expect, it } from 'vitest';
 import { AuditOutputs, auditOutputsSchema } from '@code-pushup/models';
 import { auditReport, echoRunnerConfig } from '@code-pushup/models/testing';
-import { executeProcessRunner } from './runner-process';
+import { Observer } from '@code-pushup/utils';
+import { RunnerResult, executeEsmRunner, executeProcessRunner } from './runner';
 
 const validRunnerCfg = echoRunnerConfig([auditReport()], 'output.json');
 
@@ -46,5 +47,33 @@ describe('executeRunner', () => {
     await expect(
       executeProcessRunner(runnerCfgWithErrorTransform),
     ).rejects.toThrow('transform mock error');
+  });
+});
+
+describe('executeEsmRunner', () => {
+  it('should execute valid plugin config', async () => {
+    const nextSpy = vi.fn();
+    const runnerResult: RunnerResult = await executeEsmRunner(
+      (observer?: Observer) => {
+        observer?.next?.('update');
+
+        return Promise.resolve([
+          { slug: 'mock-audit-slug', score: 0, value: 0 },
+        ] satisfies AuditOutputs);
+      },
+      { next: nextSpy },
+    );
+    expect(nextSpy).toHaveBeenCalledWith('update');
+    expect(runnerResult.audits[0]?.slug).toBe('mock-audit-slug');
+  });
+
+  it('should throw if plugin throws', async () => {
+    const nextSpy = vi.fn();
+    await expect(
+      executeEsmRunner(
+        () => Promise.reject(new Error('plugin exex mock error')),
+        { next: nextSpy },
+      ),
+    ).rejects.toThrow('plugin exex mock error');
   });
 });
